requirements.txt
Flask==1.1.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
requests==2.22.0
requests-file==1.4.3
requests-unixsocket==0.2.0
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
mysql-connector-python==8.0.25
mysqlclient==1.4.2.post1
